refactor: Move LoginUseCase into sub-package.
