+++ app/src/main/java/no/usn/mob3000/domain/usecase/auth/LoginUseCase.kt
@@ -1,4 +1,4 @@
-package no.usn.mob3000.domain.usecase
+package no.usn.mob3000.domain.usecase.auth
 
 import no.usn.mob3000.data.repository.auth.AuthRepository
 import no.usn.mob3000.data.model.auth.UserDto
